backend/translation_worker/requirements.txt
# Core dependencies for NLLB-200 Translation Worker
redis==5.0.1
aiohttp==3.9.1
aiohttp-cors==0.7.0
psutil==5.9.6
python-dotenv==1.0.0

# NLLB-200 model dependencies
transformers==4.35.2
torch==2.8.0
sentencepiece==0.2.1
protobuf==6.32.0
accelerate


# Optional: For 8-bit quantization (reduces memory usage)
# bitsandbytes==0.41.3